Move offline provider into a separate module
